% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usability.R
\name{CalcA}
\alias{CalcA}
\title{CalcA}
\usage{
CalcA(config, SDCQ, f_impression, f_amount)
}
\arguments{
\item{config}{device configuration}

\item{SDCQ}{Subjective device calibration quality}

\item{f_impression}{the direction of the first impression}

\item{f_amount}{the degree of the first impression}
}
\description{
CalcA
}
\examples{
skipped
}
