% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usability.R
\name{CalcCWL}
\alias{CalcCWL}
\title{CalcCWL}
\usage{
CalcCWL(
  pcps,
  br,
  perf,
  tct_cycle,
  nC,
  nP,
  nM,
  tct_cycle_exp_calced,
  mem_chunk_calced,
  training,
  config_DC,
  config_PR,
  config_CC
)
}
\arguments{
\item{pcps}{later...}

\item{br}{later...}

\item{perf}{later...}

\item{tct_cycle}{later...}

\item{nC}{later...}

\item{nP}{later...}

\item{nM}{later...}

\item{tct_cycle_exp_calced}{later...}

\item{mem_chunk_calced}{later...}

\item{training}{later...}

\item{config_DC}{later...}

\item{config_PR}{later...}

\item{config_CC}{later...}
}
\value{
summation of a and b
}
\description{
CalcCWL
}
\examples{
skipped
}
