% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usability.R
\name{CalcEfficiency}
\alias{CalcEfficiency}
\title{CalcEfficiency}
\usage{
CalcEfficiency(E, perf_expert)
}
\arguments{
\item{E}{error}

\item{perf_expert}{expert's performance}
}
\description{
CalcEfficiency
}
\examples{
skipped
}
