% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usability.R
\name{CalcError}
\alias{CalcError}
\title{CalcError}
\usage{
CalcError(L, output_control)
}
\arguments{
\item{L}{Learnability}

\item{output_control}{physical or virtual}
}
\description{
CalcError
}
\examples{
skipped
}
