% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usability.R
\name{CalcLearnability}
\alias{CalcLearnability}
\title{CalcLearnability}
\usage{
CalcLearnability(
  output_control,
  task_type,
  config,
  thre_min,
  thre_max,
  resil_p,
  resil_m,
  wl_p,
  wl_m,
  tct_cycle_exp_calced,
  mem_chunk_calced,
  SDCQ,
  f_impression,
  f_amount,
  pcps,
  br,
  perf,
  tct_cycle,
  nC,
  nP,
  nM,
  training,
  user_input
)
}
\arguments{
\item{output_control}{need to determine between physical and virtual}

\item{task_type}{CRT or SHAP}

\item{config}{device configuration}

\item{thre_min}{minimum threshold}

\item{thre_max}{maximum threshold}

\item{resil_p}{physical resilience}

\item{resil_m}{mental resilience}

\item{wl_p}{physical workload}

\item{wl_m}{mental workload}

\item{tct_cycle_exp_calced}{task performance one cycle calculated from the model}

\item{mem_chunk_calced}{memory chunk calculated from the model}

\item{SDCQ}{subjective device calibration quality}

\item{f_impression}{the direction of the first impression}

\item{f_amount}{the amount of the impression}

\item{pcps}{percent change of pupil size}

\item{br}{blink rate}

\item{perf}{task performance overall}

\item{tct_cycle}{tct cycle}

\item{nC}{number of cognitive operators}

\item{nP}{number of perceptual operators}

\item{nM}{number of motor operators}

\item{training}{number of required training trials}

\item{user_input}{Ask whether to use calculated value or manually added value}
}
\description{
CalcLearnability
}
\examples{
skipped
}
