% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usability.R
\name{CalcSatisfaction}
\alias{CalcSatisfaction}
\title{CalcSatisfaction}
\usage{
CalcSatisfaction(
  L,
  Effi,
  thre_min,
  thre_max,
  resil_p,
  resil_m,
  wl_p,
  wl_m,
  output_control
)
}
\arguments{
\item{L}{learnability}

\item{Effi}{efficiency}

\item{thre_min}{minimum threshold}

\item{thre_max}{maximum threshold}

\item{resil_p}{slope for p}

\item{resil_m}{slope for m}

\item{wl_p}{intercept for p}

\item{wl_m}{interecept for m}

\item{output_control}{physical or virtual}
}
\description{
CalcSatisfaction
}
\examples{
skipped
}
