% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usability.R
\name{CalcSlopeCurve}
\alias{CalcSlopeCurve}
\title{CalcSlopeCurve}
\usage{
CalcSlopeCurve(resil_p, resil_m, wl_p, wl_m, SDCQ)
}
\arguments{
\item{resil_p}{physical resilience}

\item{resil_m}{mental resilience}

\item{wl_p}{physical workload}

\item{wl_m}{mental workload}

\item{SDCQ}{Subjective device calibration quality}
}
\description{
CalcSlopeCurve
}
\examples{
skipped
}
