% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Usability.R
\name{FindX}
\alias{FindX}
\title{FindX}
\usage{
FindX(A, Index_X, thre_max)
}
\arguments{
\item{A}{tct for the first training trial}

\item{Index_X}{index}

\item{thre_max}{maximum threshold}
}
\value{
the trial number which is lower that the thre_max (moving average of three trials)
}
\description{
FindX
}
\examples{
skipped
}
