% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_01_Learnability.R
\name{GetA}
\alias{GetA}
\title{GetA()}
\usage{
GetA(raw_trn)
}
\arguments{
\item{raw_trn}{Training trials raw data}
}
\value{
A pair TCT for the first training trial
}
\description{
Calculate initial TCT in training session
}
\examples{
GetA(DC_mode, DC_gesture, PR_mode, PR_gesture)
}
