% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_01_Learnability.R
\name{GetCoef}
\alias{GetCoef}
\title{GetCoef()}
\usage{
GetCoef(raw_trn)
}
\value{
Coefficients
}
\description{
Calculate coefficients in the simulataneous equations to have initial values
}
\examples{
GetCoef()
}
