% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_01_Learnability.R
\name{GetCurve}
\alias{GetCurve}
\title{GetCurve()}
\usage{
GetCurve(A, slope_in)
}
\arguments{
\item{A}{Initial TCT of the training session}

\item{slope_in}{Slope value}
}
\value{
The number of required trials to be trained
}
\description{
Calculate the number of required trials to be trained based on configuration and cases
}
\examples{
GetCurve(A, slope value)
}
