% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_02_Errors.R
\name{GetErr}
\alias{GetErr}
\title{GetErr()}
\usage{
GetErr(oper_name)
}
\arguments{
\item{oper_name}{Name of the operator}
}
\value{
Error information in a list format: err_Time, err_Case, err_chunk, motor_err
}
\description{
Calculate errors
}
\examples{
GetErr(oper_name)
}
