% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_02_Errors.R
\name{GetErrTimeDur}
\alias{GetErrTimeDur}
\title{GetErrTimeDur()}
\usage{
GetErrTimeDur(op_P, op_C, op_M, gesture_table)
}
\arguments{
\item{op_P}{type of perceptual error}

\item{op_C}{type of cognitive error}

\item{op_M}{type of motor error}

\item{gesture_table}{A data frame for a specific type of gesture}
}
\value{
Error duration
}
\description{
Calculate the duration of an error
}
\examples{
GetErrTimeDur(op_P, op_C, op_M, gesture_table)
}
