% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_01_Learnability.R
\name{GetLearnability}
\alias{GetLearnability}
\title{GetLearnability()}
\usage{
GetLearnability(
  raw_trn,
  hook_calib_qt,
  DC_mode,
  DC_gesture,
  PR_mode,
  PR_gesture,
  slope_in
)
}
\arguments{
\item{raw_trn}{Training trials raw data}

\item{hook_calib_qt}{Case 1, 2, or 3}

\item{DC_mode}{The number of modes in DC}

\item{DC_gesture}{The number of gestures in DC}

\item{PR_mode}{The number of modes in PR}

\item{PR_gesture}{The number of gestures in PR}

\item{slope_in}{Slope value}
}
\value{
The number of required trials to be trained
}
\description{
Calculate the number of required trials to be trained
}
\examples{
GetLearnability(raw_trn, hook_calib_qt, DC_mode, DC_gesture, PR_mode, PR_gesture, slope)
}
