% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_01_Learnability.R
\name{GetMovAvg}
\alias{GetMovAvg}
\title{GetMovAvg()}
\usage{
GetMovAvg(raw_trn)
}
\arguments{
\item{raw_trn}{Training trials raw data}
}
\value{
averaged TCT for training trials
}
\description{
Calculate averages of three sequential trials
}
\examples{
GetMovAvg(raw_trn)
}
