% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_05_Satisfaction.R
\name{GetPcvdPerf}
\alias{GetPcvdPerf}
\title{GetPcvdPerf()}
\usage{
GetPcvdPerf(Learn, num_of_pins)
}
\arguments{
\item{Learn}{Not used}

\item{num_of_pins}{Number of moved pins}
}
\value{
Expected number of moved pins in the actual trial
}
\description{
Expectation on performance. Calculate the expected performance in actual trial
}
\examples{
GetPcvdPerf(LB, pin_Num)
}
