% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWL.R
\name{GetPredCRT}
\alias{GetPredCRT}
\title{GetPredCRT}
\usage{
GetPredCRT(
  pcps,
  blink,
  perf,
  tct_cycle,
  nC,
  nP,
  nM,
  tct_CPM,
  mem_load,
  training,
  config_DC,
  config_PR
)
}
\arguments{
\item{pcps}{pcps}

\item{blink}{blink}

\item{perf}{perf}

\item{tct_cycle}{tct_cycle}

\item{nC}{nC}

\item{nP}{nP}

\item{nM}{n<}

\item{tct_CPM}{tct_CPM}

\item{mem_load}{mem_load}

\item{training}{training}

\item{config_DC}{config_DC}

\item{config_PR}{config_PR}
}
\description{
GetPredCRT
}
