% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWL.R
\name{GetPredict}
\alias{GetPredict}
\title{Title GetPredict}
\usage{
GetPredict(
  pcps,
  blink,
  perf,
  tct_cycle,
  nC,
  nP,
  nM,
  tct_CPM,
  mem_load,
  training,
  config,
  task_type
)
}
\arguments{
\item{pcps}{pupil change of percentage size}

\item{blink}{blink rate}

\item{perf}{task performance}

\item{tct_cycle}{one cycle task performance}

\item{nC}{number of cognitive operators}

\item{nP}{number of perceptual operators}

\item{nM}{number of motor operators}

\item{tct_CPM}{TCT from CPM}

\item{mem_load}{memory load}

\item{training}{number of required training trials}

\item{config}{DC, PR, or CC}

\item{task_type}{CRT or SHAP}
}
\description{
Title GetPredict
}
\examples{
skipped
}
