% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_03_Memorability.R
\name{GetProbRecall}
\alias{GetProbRecall}
\title{GetProbRecall()}
\usage{
GetProbRecall(m_value)
}
\arguments{
\item{m_value}{Activation}
}
\value{
Recall probability
}
\description{
Calculate recallability using activation
}
\examples{
GetProbRecall(activation)
}
