% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_05_Satisfaction.R
\name{GetSatisfaction}
\alias{GetSatisfaction}
\title{GetSatisfaction()}
\usage{
GetSatisfaction(per_perf, desire, wts_perf, wts_desire)
}
\arguments{
\item{per_perf}{Expected number of moved pins in the actual trial}

\item{desire}{The average of desire}

\item{wts_perf}{Weight of perceived performance}

\item{wts_desire}{Weight of desire}
}
\value{

}
\description{
Calculate the weighted average of satisfaction
}
\examples{
GetSatisfaction(per_perf, desire, wts_perf, wts_desire)
}
