% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_01_Learnability.R
\name{GetUptick}
\alias{GetUptick}
\title{GetUptick()}
\usage{
GetUptick(curve_TCT)
}
\arguments{
\item{curve_TCT}{}
}
\value{
The number of required trials to be trained in Case 3
}
\description{
Calculate upticks
}
\examples{
GetUptick(curve_TCT)
}
