% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_GUI.R
\name{HPMPD}
\alias{HPMPD}
\title{HPMPD()}
\usage{
HPMPD()
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\value{
GUI

contents for ui
}
\description{
All GUI code
}
\examples{
NCPMGUI()
Shiny Server

no example
}
