% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_06_CognitiveWorkload.R
\name{lasso_pred_Result}
\alias{lasso_pred_Result}
\title{lasso_pred_Result()}
\usage{
lasso_pred_Result(lasso_prob_table, ML_outcome, acc_lasso)
}
\arguments{
\item{lasso_prob_table}{Summary table}
}
\value{
The class with highest probability
}
\description{
Extract the highest probability
}
\examples{
lasso_pred_Result(lasso_prob_table)
}
