% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_06_CognitiveWorkload.R
\name{lasso_prob}
\alias{lasso_prob}
\title{lasso_prob()}
\usage{
lasso_prob(pr_High, pr_Low, pr_Moderate, pr_Very_low, lss_summ)
}
\arguments{
\item{pr_High}{Probability of "High"}

\item{pr_Low}{Probability of "Low"}

\item{pr_Moderate}{Probability of "Moderate"}

\item{pr_Very_low}{Probability of "Very low"}

\item{lss_summ}{Summary table}
}
\value{
Probability table for each class
}
\description{
Calculate the probability of each class
}
\examples{
lasso_prob(pr_High, pr_Low, pr_Moderate, pr_Very_low, lss_summ)
}
