% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_06_CognitiveWorkload.R
\name{lasso_Summary}
\alias{lasso_Summary}
\title{lasso_Summary()}
\usage{
lasso_Summary()
}
\value{
An empty data frame
}
\description{
Create a data table for LASSO summary
}
\examples{
lasso_Summary()
}
