% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part_GUI.R
\name{NewOper}
\alias{NewOper}
\title{NewOper}
\usage{
NewOper(op_name, op_time, oper_set_new)
}
\arguments{
\item{op_name}{operator name}

\item{op_time}{operator time}
}
\description{
NewOper
}
\examples{
NewOper(a, b)
}
