% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_HPM_UP.R
\name{RunHPMUP}
\alias{RunHPMUP}
\title{Title HPM-UP main function}
\usage{
RunHPMUP(scenario, time_library)
}
\arguments{
\item{scenario}{: is a scenario file (csv format)}
}
\value{
All results
}
\description{
Title HPM-UP main function
}
\examples{
RunHPMUP(scenario)
}
