% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/u_04_Efficiency.R
\name{Shiny_Parameter_Test}
\alias{Shiny_Parameter_Test}
\title{Shiny_Parameter_Test()}
\usage{
Shiny_Parameter_Test(test)
}
\arguments{
\item{test}{}
}
\value{
add + 5
}
\description{
testing parameter calculation between R and Shiny
}
