% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_difficulty.R
\name{TaskDiff}
\alias{TaskDiff}
\title{TaskDiff()}
\usage{
TaskDiff(line_of_input)
}
\arguments{
\item{line_of_input}{}
}
\value{
operator name
}
\description{
Return the number of cognitive operators
}
\examples{
ExtOper(line_of_input)
}
