% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_HPM_UP.R
\name{test_Shiny_Inputs}
\alias{test_Shiny_Inputs}
\title{test_Shiny_Inputs}
\usage{
test_Shiny_Inputs(config, thre, theta, cq, lot, resil_p, resil_m, wl_p, wl_m)
}
\arguments{
\item{wl_m}{}
}
\value{

}
\description{
test_Shiny_Inputs
}
