% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_GUI.R
\name{USABILATOR_calc}
\alias{USABILATOR_calc}
\title{USABILATOR_calc}
\usage{
USABILATOR_calc(scenario, time_library)
}
\arguments{
\item{scenario}{scenario file}
}
\value{
all outcome
}
\description{
Runs N-CPM model
}
\examples{
USABILATOR_calc(sce, "Novice")
}
