% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_GUI.R
\name{USABILATOR_GUI}
\alias{USABILATOR_GUI}
\title{USABILATOR_GUI()}
\usage{
USABILATOR_GUI()
}
\arguments{
\item{input}{input}

\item{output}{output}

\item{session}{session}
}
\value{
GUI

contents for ui
}
\description{
All GUI code
}
\examples{
USABILATOR_GUI()
Shiny Server

no example
}
